+++ lib/nothing_glyph_interface_method_channel.dart
@@ -20,8 +20,7 @@ class MethodChannelNothingGlyphInterface extends NothingGlyphInterfacePlatform {
         default:
           throw PlatformException(
             code: 'Unimplemented',
-            details:
-                "The glyph_interface plugin for Flutter does not implement the method '${call.method}'",
+            details: "The glyph_interface plugin for Flutter does not implement the method '${call.method}'",
           );
       }
     });
@@ -29,8 +28,7 @@ class MethodChannelNothingGlyphInterface extends NothingGlyphInterfacePlatform {
 
   @override
   Future<String?> getPlatformVersion() async {
-    final version =
-        await methodChannel.invokeMethod<String>('getPlatformVersion');
+    final version = await methodChannel.invokeMethod<String>('getPlatformVersion');
     return version;
   }
 
@@ -49,6 +47,11 @@ class MethodChannelNothingGlyphInterface extends NothingGlyphInterfacePlatform {
   @override
   Future<bool?> is23111() async {
     return await methodChannel.invokeMethod<bool>('is23111');
+  }
+
+  @override
+  Future<bool?> is23113() async {
+    return await methodChannel.invokeMethod<bool>('is23113');
   }
 
   // Gylph Frame
@@ -80,8 +83,7 @@ class MethodChannelNothingGlyphInterface extends NothingGlyphInterfacePlatform {
 
   @override
   Future<void> buildGlyphFrame(List<Map<String, int?>> operations) async {
-    await methodChannel
-        .invokeMethod('buildGlyphFrame', {'operations': operations});
+    await methodChannel.invokeMethod('buildGlyphFrame', {'operations': operations});
   }
 
   @override
@@ -96,15 +98,12 @@ class MethodChannelNothingGlyphInterface extends NothingGlyphInterfacePlatform {
 
   @override
   Future<void> displayProgress(int progress, {bool reverse = false}) async {
-    await methodChannel.invokeMethod(
-        'displayProgress', {'progress': progress, 'reverse': reverse});
+    await methodChannel.invokeMethod('displayProgress', {'progress': progress, 'reverse': reverse});
   }
 
   @override
-  Future<void> displayProgressAndToggle(int progress,
-      {bool reverse = false}) async {
-    await methodChannel.invokeMethod(
-        'displayProgressAndToggle', {'progress': progress, 'reverse': reverse});
+  Future<void> displayProgressAndToggle(int progress, {bool reverse = false}) async {
+    await methodChannel.invokeMethod('displayProgressAndToggle', {'progress': progress, 'reverse': reverse});
   }
 
   @override